+++ guidance/reference_filter_dp/CMakeLists.txt
@@ -24,19 +24,34 @@ find_package(fmt REQUIRED)
 
 include_directories(include)
 
-set(LIB_NAME "${PROJECT_NAME}_component")
+set(CORE_LIB_NAME "${PROJECT_NAME}")
 
-add_library(${LIB_NAME} SHARED
-  src/reference_filter.cpp
-  src/reference_filter_ros.cpp)
+add_library(${CORE_LIB_NAME} SHARED
+  src/lib/reference_filter.cpp
+)
 
-ament_target_dependencies(${LIB_NAME} PUBLIC
+target_include_directories(${CORE_LIB_NAME}
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:include>
+)
+
+ament_target_dependencies(${CORE_LIB_NAME}
+  Eigen3
+)
+
+set(COMPONENT_LIB_NAME "${PROJECT_NAME}_component")
+
+add_library(${COMPONENT_LIB_NAME} SHARED
+  src/ros/reference_filter_ros.cpp
+)
+
+ament_target_dependencies(${COMPONENT_LIB_NAME}
   rclcpp
   rclcpp_components
   rclcpp_action
   geometry_msgs
   nav_msgs
-  Eigen3
   vortex_msgs
   vortex_utils
   vortex_utils_ros
@@ -44,16 +59,19 @@ ament_target_dependencies(${LIB_NAME} PUBLIC
   fmt
 )
 
+target_link_libraries(${COMPONENT_LIB_NAME} ${CORE_LIB_NAME})
+
 rclcpp_components_register_node(
-  ${LIB_NAME}
+  ${COMPONENT_LIB_NAME}
   PLUGIN "ReferenceFilterNode"
   EXECUTABLE ${PROJECT_NAME}_node
 )
 
-ament_export_targets(export_${LIB_NAME})
-
-install(TARGETS ${LIB_NAME}
-  EXPORT export_${LIB_NAME}
+install(
+  TARGETS
+    ${CORE_LIB_NAME}
+    ${COMPONENT_LIB_NAME}
+  EXPORT export_${PROJECT_NAME}
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
   RUNTIME DESTINATION bin
@@ -69,6 +87,13 @@ install(DIRECTORY
   config
   DESTINATION share/${PROJECT_NAME}/
 )
+
+ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
+ament_export_dependencies(
+  Eigen3
+)
+ament_export_include_directories(include)
+ament_export_libraries(${CORE_LIB_NAME})
 
 if(BUILD_TESTING)
   add_subdirectory(test)
