+++ guidance/reference_filter_dp/include/reference_filter_dp/lib/reference_filter.hpp
@@ -1,7 +1,7 @@
-#ifndef REFERENCE_FILTER_DP__REFERENCE_FILTER_HPP_
-#define REFERENCE_FILTER_DP__REFERENCE_FILTER_HPP_
+#ifndef REFERENCE_FILTER_DP__LIB__REFERENCE_FILTER_HPP_
+#define REFERENCE_FILTER_DP__LIB__REFERENCE_FILTER_HPP_
 
-#include "reference_filter_dp/eigen_typedefs.hpp"
+#include "reference_filter_dp/lib/eigen_typedefs.hpp"
 
 namespace vortex::guidance {
 
@@ -41,4 +41,4 @@ class ReferenceFilter {
 
 }  // namespace vortex::guidance
 
-#endif  // REFERENCE_FILTER_DP__REFERENCE_FILTER_HPP_
+#endif  // REFERENCE_FILTER_DP__LIB__REFERENCE_FILTER_HPP_
