+++ guidance/reference_filter_dp/src/ros/reference_filter_ros.cpp
@@ -1,4 +1,4 @@
-#include "reference_filter_dp/reference_filter_ros.hpp"
+#include "reference_filter_dp/ros/reference_filter_ros.hpp"
 #include <spdlog/spdlog.h>
 #include <mutex>
 #include <rclcpp_components/register_node_macro.hpp>
