+++ guidance/reference_filter_dp/test/test_reference_filter.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
-#include "reference_filter_dp/eigen_typedefs.hpp"
-#include "reference_filter_dp/reference_filter.hpp"
+#include "reference_filter_dp/lib/eigen_typedefs.hpp"
+#include "reference_filter_dp/lib/reference_filter.hpp"
 
 namespace vortex::guidance {
